Moar comments/javadoc

+++ src/fr/epf/computer/controller/AddComputerController.java
@@ -36,6 +36,7 @@ public class AddComputerController extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
         // Boolean controlling whether all form inputs are valid
+        // Will be set to "false" if things go bad
         boolean bEverythingOkay = true;
 
         // Computer name
@@ -54,13 +55,18 @@ public class AddComputerController extends HttpServlet {
         // Storing given value in case something goes wrong and we need to display it back to the user
         request.setAttribute("dateIntroducedValue", recup);
 
+        // Human readable format: YYYY-MM-DD (Year-Month-Date)
     	SimpleDateFormat sdf = new SimpleDateFormat("y-MM-dd");
+
+        // SDF, don't be lenient, bro.
+        // SDF should only parse what it's told to
         sdf.setLenient(false);
-    	Date introduced = null;
+
+        Date introduced = null;
     	try{
     		introduced = sdf.parse(recup);
     	}catch(ParseException e){
-    		
+
     	}
         if(introduced == null) {
             // Error: date introduced is invalid
@@ -92,7 +98,7 @@ public class AddComputerController extends HttpServlet {
 
         Company company = null;
     	try{
-    		int companyId = Integer.parseInt(companyRecup);
+            long companyId = Long.parseLong(companyRecup);
             company = companyService.findById(companyId);
     	}catch(NumberFormatException e){
 
@@ -104,15 +110,21 @@ public class AddComputerController extends HttpServlet {
         }
 
         if( bEverythingOkay ) {
+            // Everything went according to plan, so we can build a full instance of Computer
+            // ComputerService will take care of persisting that new instance
             computerService.create( new Computer.Builder()
                     .name(name)
                     .introduced(introduced)
                     .discontinued(discontinued)
                     .company(company)
                     .build());
+
+            // All done, let's go back to the list
             response.sendRedirect("/computer/list");
         }
         else {
+            // Tough luck, user didn't fill the form with valid input
+            // Let's just redirect him back to the form and show him what he did wrong
             request.setAttribute("bEverythingOkay", false);
             doGet(request, response);
         }
@@ -120,10 +132,10 @@ public class AddComputerController extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         
-    	//Companies list send in request :
-    	request.setAttribute("companies", companyService.getCompanies());
-    	
-    	RequestDispatcher rd = getServletContext().getRequestDispatcher("/WEB-INF/add_computer.jsp");
+        //Companies list send in request :
+        request.setAttribute("companies", companyService.getCompanies());
+
+        RequestDispatcher rd = getServletContext().getRequestDispatcher("/WEB-INF/add_computer.jsp");
         rd.forward(request, response);
     }
 }
